Initial setup of line chart
